+++ Source/CesiumRuntime/CesiumRuntime.Build.cs
@@ -65,7 +65,7 @@ public class CesiumRuntime : ModuleRules
         string[] libs = new string[]
         {
             "async++",
-            "Cesium3DTiles",
+            "Cesium3DTilesSelection",
             "CesiumAsync",
             "CesiumGeometry",
             "CesiumGeospatial",