+++ Source/CesiumRuntime/Private/Cesium3DTileset.cpp
@@ -2,14 +2,14 @@
 
 #include "Cesium3DTileset.h"
 #include "Camera/PlayerCameraManager.h"
-#include "Cesium3DTiles/BingMapsRasterOverlay.h"
-#include "Cesium3DTiles/CartographicSelection.h"
-#include "Cesium3DTiles/CreditSystem.h"
-#include "Cesium3DTiles/GltfContent.h"
-#include "Cesium3DTiles/IPrepareRendererResources.h"
-#include "Cesium3DTiles/RasterizedPolygonsOverlay.h"
-#include "Cesium3DTiles/Tileset.h"
-#include "Cesium3DTiles/TilesetOptions.h"
+#include "Cesium3DTilesSelection/BingMapsRasterOverlay.h"
+#include "Cesium3DTilesSelection/CartographicSelection.h"
+#include "Cesium3DTilesSelection/CreditSystem.h"
+#include "Cesium3DTilesSelection/GltfContent.h"
+#include "Cesium3DTilesSelection/IPrepareRendererResources.h"
+#include "Cesium3DTilesSelection/RasterizedPolygonsOverlay.h"
+#include "Cesium3DTilesSelection/Tileset.h"
+#include "Cesium3DTilesSelection/TilesetOptions.h"
 #include "Cesium3DTilesetRoot.h"
 #include "CesiumAsync/CachingAssetAccessor.h"
 #include "CesiumAsync/SqliteCache.h"
@@ -261,12 +261,13 @@ void ACesium3DTileset::OnFocusEditorViewportOnThis() {
     }
   };
 
-  const Cesium3DTiles::Tile* pRootTile = this->_pTileset->getRootTile();
+  const Cesium3DTilesSelection::Tile* pRootTile =
+      this->_pTileset->getRootTile();
   if (!pRootTile) {
     return;
   }
 
-  const Cesium3DTiles::BoundingVolume& boundingVolume =
+  const Cesium3DTilesSelection::BoundingVolume& boundingVolume =
       pRootTile->getBoundingVolume();
 
   // calculate unreal camera position
@@ -279,7 +280,7 @@ void ACesium3DTileset::OnFocusEditorViewportOnThis() {
 
   // calculate unreal camera orientation
   glm::dvec3 ecefCenter =
-      Cesium3DTiles::getBoundingVolumeCenter(boundingVolume);
+      Cesium3DTilesSelection::getBoundingVolumeCenter(boundingVolume);
   glm::dvec3 unrealCenter = transform * glm::dvec4(ecefCenter, 1.0);
   glm::dvec3 unrealCameraFront =
       glm::normalize(unrealCenter - unrealCameraPosition);
@@ -331,7 +332,7 @@ bool ACesium3DTileset::IsBoundingVolumeReady() const {
   return this->_pTileset && this->_pTileset->getRootTile();
 }
 
-std::optional<Cesium3DTiles::BoundingVolume>
+std::optional<Cesium3DTilesSelection::BoundingVolume>
 ACesium3DTileset::GetBoundingVolume() const {
   if (!this->IsBoundingVolumeReady()) {
     return std::nullopt;
@@ -433,7 +434,8 @@ void ACesium3DTileset::NotifyHit(
   // std::cout << "Hit face index 2: " << detailedHit.FaceIndex << std::endl;
 }
 
-class UnrealResourcePreparer : public Cesium3DTiles::IPrepareRendererResources {
+class UnrealResourcePreparer
+    : public Cesium3DTilesSelection::IPrepareRendererResources {
 public:
   UnrealResourcePreparer(ACesium3DTileset* pActor)
       : _pActor(pActor)
@@ -461,9 +463,10 @@ public:
   }
 
   virtual void* prepareInMainThread(
-      Cesium3DTiles::Tile& tile,
+      Cesium3DTilesSelection::Tile& tile,
       void* pLoadThreadResult) override {
-    const Cesium3DTiles::TileContentLoadResult* pContent = tile.getContent();
+    const Cesium3DTilesSelection::TileContentLoadResult* pContent =
+        tile.getContent();
     if (pContent && pContent->model) {
       std::unique_ptr<UCesiumGltfComponent::HalfConstructed> pHalf(
           reinterpret_cast<UCesiumGltfComponent::HalfConstructed*>(
@@ -481,7 +484,7 @@ public:
   }
 
   virtual void free(
-      Cesium3DTiles::Tile& tile,
+      Cesium3DTilesSelection::Tile& tile,
       void* pLoadThreadResult,
       void* pMainThreadResult) noexcept override {
     if (pLoadThreadResult) {
@@ -502,7 +505,7 @@ public:
   }
 
   virtual void* prepareRasterInMainThread(
-      const Cesium3DTiles::RasterOverlayTile& rasterTile,
+      const Cesium3DTilesSelection::RasterOverlayTile& rasterTile,
       void* pLoadThreadResult) override {
     const CesiumGltf::ImageCesium& image = rasterTile.getImage();
     if (image.width <= 0 || image.height <= 0) {
@@ -543,7 +546,7 @@ public:
   }
 
   virtual void freeRaster(
-      const Cesium3DTiles::RasterOverlayTile& rasterTile,
+      const Cesium3DTilesSelection::RasterOverlayTile& rasterTile,
       void* pLoadThreadResult,
       void* pMainThreadResult) noexcept override {
     UTexture2D* pTexture = static_cast<UTexture2D*>(pMainThreadResult);
@@ -555,14 +558,15 @@ public:
   }
 
   virtual void attachRasterInMainThread(
-      const Cesium3DTiles::Tile& tile,
+      const Cesium3DTilesSelection::Tile& tile,
       uint32_t overlayTextureCoordinateID,
-      const Cesium3DTiles::RasterOverlayTile& rasterTile,
+      const Cesium3DTilesSelection::RasterOverlayTile& rasterTile,
       void* pMainThreadRendererResources,
       const CesiumGeometry::Rectangle& textureCoordinateRectangle,
       const glm::dvec2& translation,
       const glm::dvec2& scale) override {
-    const Cesium3DTiles::TileContentLoadResult* pContent = tile.getContent();
+    const Cesium3DTilesSelection::TileContentLoadResult* pContent =
+        tile.getContent();
     if (pContent && pContent->model) {
       UCesiumGltfComponent* pGltfContent =
           reinterpret_cast<UCesiumGltfComponent*>(tile.getRendererResources());
@@ -579,13 +583,14 @@ public:
   }
 
   virtual void detachRasterInMainThread(
-      const Cesium3DTiles::Tile& tile,
+      const Cesium3DTilesSelection::Tile& tile,
       uint32_t overlayTextureCoordinateID,
-      const Cesium3DTiles::RasterOverlayTile& rasterTile,
+      const Cesium3DTilesSelection::RasterOverlayTile& rasterTile,
       void* pMainThreadRendererResources,
       const CesiumGeometry::Rectangle& textureCoordinateRectangle) noexcept
       override {
-    const Cesium3DTiles::TileContentLoadResult* pContent = tile.getContent();
+    const Cesium3DTilesSelection::TileContentLoadResult* pContent =
+        tile.getContent();
     if (pContent && pContent->model) {
       UCesiumGltfComponent* pGltfContent =
           reinterpret_cast<UCesiumGltfComponent*>(tile.getRendererResources());
@@ -683,7 +688,7 @@ void ACesium3DTileset::LoadTileset() {
     this->CreditSystem = ACesiumCreditSystem::GetDefaultForActor(this);
   }
 
-  Cesium3DTiles::TilesetExternals externals{
+  Cesium3DTilesSelection::TilesetExternals externals{
       pAssetAccessor,
       std::make_shared<UnrealResourcePreparer>(this),
       asyncSystem,
@@ -693,7 +698,7 @@ void ACesium3DTileset::LoadTileset() {
 
   this->_startTime = std::chrono::high_resolution_clock::now();
 
-  Cesium3DTiles::TilesetOptions options;
+  Cesium3DTilesSelection::TilesetOptions options;
   /*
   for (ACesiumCartographicSelection* pCartographicSelection :
        this->CartographicSelections) {
@@ -712,7 +717,7 @@ void ACesium3DTileset::LoadTileset() {
   switch (this->TilesetSource) {
   case ETilesetSource::FromUrl:
     UE_LOG(LogCesium, Log, TEXT("Loading tileset from URL %s"), *this->Url);
-    this->_pTileset = new Cesium3DTiles::Tileset(
+    this->_pTileset = new Cesium3DTilesSelection::Tileset(
         externals,
         TCHAR_TO_UTF8(*this->Url),
         options);
@@ -723,7 +728,7 @@ void ACesium3DTileset::LoadTileset() {
         Log,
         TEXT("Loading tileset for asset ID %d"),
         this->IonAssetID);
-    this->_pTileset = new Cesium3DTiles::Tileset(
+    this->_pTileset = new Cesium3DTilesSelection::Tileset(
         externals,
         static_cast<uint32_t>(this->IonAssetID),
         TCHAR_TO_UTF8(*this->IonAccessToken),
@@ -741,7 +746,7 @@ void ACesium3DTileset::LoadTileset() {
     }
 
     FString textureName = pCartographicSelection->TargetTexture;
-    std::vector<Cesium3DTiles::CartographicSelection> polygons;
+    std::vector<Cesium3DTilesSelection::CartographicSelection> polygons;
 
     for (int j = i; j < this->CartographicSelections.Num(); ++j) {
       ACesiumCartographicSelection* pCartographicSelection2 =
@@ -757,7 +762,7 @@ void ACesium3DTileset::LoadTileset() {
     }
 
     this->_pTileset->getOverlays().add(
-        std::make_unique<Cesium3DTiles::RasterizedPolygonsOverlay>(
+        std::make_unique<Cesium3DTilesSelection::RasterizedPolygonsOverlay>(
             TCHAR_TO_UTF8(*textureName),
             polygons,
             CesiumGeospatial::Ellipsoid::WGS84,
@@ -879,11 +884,16 @@ ACesium3DTileset::GetPlayerCameras() const {
     pStereoRendering = GEngine->StereoRenderingDevice;
   }
 
+  bool useStereoRendering = false;
+  if (pStereoRendering && pStereoRendering->IsStereoEnabled()) {
+    useStereoRendering = true;
+  }
+
   uint32 stereoLeftSizeX = static_cast<uint32>(size.X);
   uint32 stereoLeftSizeY = static_cast<uint32>(size.Y);
   uint32 stereoRightSizeX = static_cast<uint32>(size.X);
   uint32 stereoRightSizeY = static_cast<uint32>(size.Y);
-  if (pStereoRendering) {
+  if (useStereoRendering) {
     int32 _x;
     int32 _y;
 
@@ -903,9 +913,7 @@ ACesium3DTileset::GetPlayerCameras() const {
   }
 
   FVector2D stereoLeftSize(stereoLeftSizeX, stereoRightSizeY);
-  float stereoLeftAspectRatio = stereoLeftSizeY / stereoLeftSizeX;
   FVector2D stereoRightSize(stereoRightSizeX, stereoRightSizeY);
-  float stereoRightAspectRatio = stereoRightSizeY / stereoRightSizeX;
 
   std::vector<ACesium3DTileset::UnrealCameraParameters> cameras;
   cameras.reserve(pWorld->GetNumPlayerControllers());
@@ -933,50 +941,55 @@ ACesium3DTileset::GetPlayerCameras() const {
     FRotator rotation;
     pPlayerController->GetPlayerViewPoint(location, rotation);
 
-    if (pStereoRendering) {
-      FVector leftEyeLocation = location;
-      FRotator leftEyeRotation = rotation;
-      pStereoRendering->CalculateStereoViewOffset(
-          EStereoscopicPass::eSSP_LEFT_EYE,
-          leftEyeRotation,
-          worldToMeters,
-          leftEyeLocation);
+    if (useStereoRendering) {
+      if (stereoLeftSize.X >= 1.0 && stereoLeftSize.Y >= 1.0) {
+        FVector leftEyeLocation = location;
+        FRotator leftEyeRotation = rotation;
+        pStereoRendering->CalculateStereoViewOffset(
+            EStereoscopicPass::eSSP_LEFT_EYE,
+            leftEyeRotation,
+            worldToMeters,
+            leftEyeLocation);
 
-      FMatrix projection = pStereoRendering->GetStereoProjectionMatrix(
-          EStereoscopicPass::eSSP_LEFT_EYE);
+        FMatrix projection = pStereoRendering->GetStereoProjectionMatrix(
+            EStereoscopicPass::eSSP_LEFT_EYE);
 
-      // TODO: consider assymetric frustums using 4 fovs
-      float one_over_tan_half_hfov = projection.M[0][0];
+        // TODO: consider assymetric frustums using 4 fovs
+        float one_over_tan_half_hfov = projection.M[0][0];
 
-      float hfov =
-          glm::degrees(2.0f * glm::atan(1.0f / one_over_tan_half_hfov));
+        float hfov =
+            glm::degrees(2.0f * glm::atan(1.0f / one_over_tan_half_hfov));
 
-      cameras.push_back(UnrealCameraParameters{
-          stereoLeftSize,
-          leftEyeLocation,
-          leftEyeRotation,
-          hfov});
+        cameras.push_back(UnrealCameraParameters{
+            stereoLeftSize,
+            leftEyeLocation,
+            leftEyeRotation,
+            hfov});
+      }
 
-      FVector rightEyeLocation = location;
-      FRotator rightEyeRotation = rotation;
-      pStereoRendering->CalculateStereoViewOffset(
-          EStereoscopicPass::eSSP_RIGHT_EYE,
-          rightEyeRotation,
-          worldToMeters,
-          rightEyeLocation);
+      if (stereoRightSize.X >= 1.0 && stereoRightSize.Y >= 1.0) {
+        FVector rightEyeLocation = location;
+        FRotator rightEyeRotation = rotation;
+        pStereoRendering->CalculateStereoViewOffset(
+            EStereoscopicPass::eSSP_RIGHT_EYE,
+            rightEyeRotation,
+            worldToMeters,
+            rightEyeLocation);
 
-      projection = pStereoRendering->GetStereoProjectionMatrix(
-          EStereoscopicPass::eSSP_RIGHT_EYE);
+        FMatrix projection = pStereoRendering->GetStereoProjectionMatrix(
+            EStereoscopicPass::eSSP_RIGHT_EYE);
 
-      one_over_tan_half_hfov = projection.M[0][0];
+        float one_over_tan_half_hfov = projection.M[0][0];
 
-      hfov = glm::degrees(2.0f * glm::atan(1.0f / one_over_tan_half_hfov));
+        float hfov =
+            glm::degrees(2.0f * glm::atan(1.0f / one_over_tan_half_hfov));
 
-      cameras.push_back(UnrealCameraParameters{
-          stereoRightSize,
-          rightEyeLocation,
-          rightEyeRotation,
-          hfov});
+        cameras.push_back(UnrealCameraParameters{
+            stereoRightSize,
+            rightEyeLocation,
+            rightEyeRotation,
+            hfov});
+      }
     } else {
       cameras.push_back(UnrealCameraParameters{size, location, rotation, fov});
     }
@@ -985,7 +998,7 @@ ACesium3DTileset::GetPlayerCameras() const {
   return cameras;
 }
 
-/*static*/ Cesium3DTiles::ViewState
+/*static*/ Cesium3DTilesSelection::ViewState
 ACesium3DTileset::CreateViewStateFromViewParameters(
     const UnrealCameraParameters& camera,
     const glm::dmat4& unrealWorldToTileset) {
@@ -1008,7 +1021,7 @@ ACesium3DTileset::CreateViewStateFromViewParameters(
   glm::dvec3 tilesetCameraUp =
       glm::normalize(unrealWorldToTileset * glm::dvec4(up.X, up.Y, up.Z, 0.0));
 
-  return Cesium3DTiles::ViewState::create(
+  return Cesium3DTilesSelection::ViewState::create(
       tilesetCameraLocation,
       tilesetCameraFront,
       tilesetCameraUp,
@@ -1068,7 +1081,7 @@ namespace {
  * TODO Add details here what that means
  * Old comment:
  * Consider Exclusion zone to drop this tile... Ideally, should be
- * considered in Cesium3DTiles::ViewState to avoid loading the tile
+ * considered in Cesium3DTilesSelection::ViewState to avoid loading the tile
  * first...
  *
  * @param exclusionZones The exclusion zones
@@ -1077,7 +1090,7 @@ namespace {
  */
 bool isInExclusionZone(
     const TArray<FCesiumExclusionZone>& exclusionZones,
-    Cesium3DTiles::Tile const* tile) {
+    Cesium3DTilesSelection::Tile const* tile) {
   if (exclusionZones.Num() == 0) {
     return false;
   }
@@ -1103,13 +1116,13 @@ bool isInExclusionZone(
 }
 
 void removeVisibleTilesFromList(
-    std::vector<Cesium3DTiles::Tile*>& list,
-    const std::vector<Cesium3DTiles::Tile*>& visibleTiles) {
+    std::vector<Cesium3DTilesSelection::Tile*>& list,
+    const std::vector<Cesium3DTilesSelection::Tile*>& visibleTiles) {
   if (list.empty()) {
     return;
   }
 
-  for (Cesium3DTiles::Tile* pTile : visibleTiles) {
+  for (Cesium3DTilesSelection::Tile* pTile : visibleTiles) {
     auto it = std::find(list.begin(), list.end(), pTile);
     if (it != list.end()) {
       list.erase(it);
@@ -1126,9 +1139,10 @@ void removeVisibleTilesFromList(
  *
  * @param tiles The tiles to hide
  */
-void hideTilesToNoLongerRender(const std::vector<Cesium3DTiles::Tile*>& tiles) {
-  for (Cesium3DTiles::Tile* pTile : tiles) {
-    if (pTile->getState() != Cesium3DTiles::Tile::LoadState::Done) {
+void hideTilesToNoLongerRender(
+    const std::vector<Cesium3DTilesSelection::Tile*>& tiles) {
+  for (Cesium3DTilesSelection::Tile* pTile : tiles) {
+    if (pTile->getState() != Cesium3DTilesSelection::Tile::LoadState::Done) {
       continue;
     }
 
@@ -1178,7 +1192,8 @@ void applyActorCollisionSettings(
 } // namespace
 
 void ACesium3DTileset::updateTilesetOptionsFromProperties() {
-  Cesium3DTiles::TilesetOptions& options = this->_pTileset->getOptions();
+  Cesium3DTilesSelection::TilesetOptions& options =
+      this->_pTileset->getOptions();
   options.maximumScreenSpaceError =
       static_cast<double>(this->MaximumScreenSpaceError);
 
@@ -1196,7 +1211,7 @@ void ACesium3DTileset::updateTilesetOptionsFromProperties() {
 }
 
 void ACesium3DTileset::updateLastViewUpdateResultState(
-    const Cesium3DTiles::ViewUpdateResult& result) {
+    const Cesium3DTilesSelection::ViewUpdateResult& result) {
   if (result.tilesToRenderThisFrame.size() != this->_lastTilesRendered ||
       result.tilesLoadingLowPriority != this->_lastTilesLoadingLowPriority ||
       result.tilesLoadingMediumPriority !=
@@ -1237,9 +1252,9 @@ void ACesium3DTileset::updateLastViewUpdateResultState(
 }
 
 void ACesium3DTileset::showTilesToRender(
-    const std::vector<Cesium3DTiles::Tile*>& tiles) {
-  for (Cesium3DTiles::Tile* pTile : tiles) {
-    if (pTile->getState() != Cesium3DTiles::Tile::LoadState::Done) {
+    const std::vector<Cesium3DTilesSelection::Tile*>& tiles) {
+  for (Cesium3DTilesSelection::Tile* pTile : tiles) {
+    if (pTile->getState() != Cesium3DTilesSelection::Tile::LoadState::Done) {
       continue;
     }
 
@@ -1248,7 +1263,7 @@ void ACesium3DTileset::showTilesToRender(
     }
 
     // That looks like some reeeally entertaining debug session...:
-    // const Cesium3DTiles::TileID& id = pTile->getTileID();
+    // const Cesium3DTilesSelection::TileID& id = pTile->getTileID();
     // const CesiumGeometry::QuadtreeTileID* pQuadtreeID =
     // std::get_if<CesiumGeometry::QuadtreeTileID>(&id); if (!pQuadtreeID ||
     // pQuadtreeID->level != 14 || pQuadtreeID->x != 5503 || pQuadtreeID->y !=
@@ -1275,9 +1290,10 @@ void ACesium3DTileset::showTilesToRender(
           this->RootComponent,
           FAttachmentTransformRules::KeepRelativeTransform);
       if (!attached) {
-        FString tileIdString(Cesium3DTiles::TileIdUtilities::createTileIdString(
-                                 pTile->getTileID())
-                                 .c_str());
+        FString tileIdString(
+            Cesium3DTilesSelection::TileIdUtilities::createTileIdString(
+                pTile->getTileID())
+                .c_str());
         UE_LOG(
             LogCesium,
             Warning,
@@ -1345,13 +1361,13 @@ void ACesium3DTileset::Tick(float DeltaTime) {
   glm::dmat4 unrealWorldToTileset = glm::affineInverse(
       this->GetCesiumTilesetToUnrealRelativeWorldTransform());
 
-  std::vector<Cesium3DTiles::ViewState> frustums;
+  std::vector<Cesium3DTilesSelection::ViewState> frustums;
   for (const UnrealCameraParameters& camera : cameras) {
     frustums.push_back(
         CreateViewStateFromViewParameters(camera, unrealWorldToTileset));
   }
 
-  const Cesium3DTiles::ViewUpdateResult& result =
+  const Cesium3DTilesSelection::ViewUpdateResult& result =
       this->_captureMovieMode ? this->_pTileset->updateViewOffline(frustums)
                               : this->_pTileset->updateView(frustums);
   updateLastViewUpdateResultState(result);