+++ Source/CesiumRuntime/Public/CesiumBingMapsRasterOverlay.h
@@ -47,6 +47,6 @@ public:
   EBingMapsStyle MapStyle = EBingMapsStyle::Aerial;
 
 protected:
-  virtual std::unique_ptr<Cesium3DTiles::RasterOverlay>
+  virtual std::unique_ptr<Cesium3DTilesSelection::RasterOverlay>
   CreateOverlay() override;
 };