+++ CHANGES.md
@@ -1,16 +1,22 @@
 # Change Log
 
-### Next Release - ?
+### v1.5.1 - 2021-08-09
+
+##### Breaking :mega:
+
+- Changed Cesium Native Cesium3DTiles's namespace to Cesium3DTilesSelection's namespace
 
 ##### Fixes :wrench:
 
 - Fixed a bug that could cause mis-registration of feature metadata to the wrong features in Draco-compressed meshes.
+- Fixed a bug that could cause a crash with VR/AR devices enabled but not in use.
 
 ### v1.5.0 - 2021-08-02
 
 ##### Additions :tada:
 
 - Added support for reading per-feature metadata from glTFs with the `EXT_feature_metadata` extension or from 3D Tiles with a B3DM batch table and accessing it from Blueprints.
+- Added support for using multiple view frustums in `ACesium3DTileset` to inform the tile selection algorithm.
 
 ##### Fixes :wrench: